+++ laktory/models/group.py
@@ -12,6 +12,6 @@ class Group(BaseModel, Resources):
     # Resources Engine Methods                                                #
     # ----------------------------------------------------------------------- #
 
-    def deploy_with_pulumi(self, name=None, **kwargs):
+    def deploy_with_pulumi(self, name=None, opts=None):
         from laktory.resourcesengines.pulumi.group import PulumiGroup
-        return PulumiGroup(name=name, group=self, **kwargs)
+        return PulumiGroup(name=name, group=self, opts=opts)