+++ laktory/resourcesengines/pulumi/group.py
@@ -15,19 +15,19 @@ class PulumiGroup(PulumiResourcesEngine):
             name=None,
             group: Group = None,
             opts=None,
-            **kwargs
     ):
         if name is None:
             name = f"group-{group.display_name}"
         super().__init__(self.t, name, {}, opts)
 
-        kwargs["opts"] = kwargs.get("opts", pulumi.ResourceOptions())
-        kwargs["opts"].parent = self
-        kwargs["opts"].delete_before_replace = getattr(kwargs["opts"], "delete_before_replace", True)
+        opts = pulumi.ResourceOptions(
+            parent=self,
+            delete_before_replace=True,
+        )
 
         self.group = databricks.Group(
             f"group-{group.display_name}",
             display_name=group.display_name,
             allow_cluster_create=group.allow_cluster_create,
-            **kwargs,
+            opts=opts,
         )