+++ laktory/resourcesengines/pulumi/serviceprincipal.py
@@ -18,41 +18,38 @@ class PulumiServicePrincipal(PulumiResourcesEngine):
     def provider(self):
         return "databricks"
 
-    @property
-    def default_name(self) -> str:
-        return f"service-principal-{self.model.display_name}"
-
     def __init__(
             self,
             name=None,
             service_principal: ServicePrincipal = None,
             groups: Union[list[Group], dict] = None,
             opts=None,
-            **kwargs
     ):
         sp = service_principal
         if name is None:
             name = f"service-principal-{sp.display_name}"
         super().__init__(self.t, name, {}, opts)
 
-        kwargs["opts"] = kwargs.get("opts", pulumi.ResourceOptions())
-        kwargs["opts"].parent = self
-        kwargs["opts"].delete_before_replace = getattr(kwargs["opts"], "delete_before_replace", True)
+        opts = pulumi.ResourceOptions(
+            parent=self,
+            delete_before_replace=True,
+        )
 
         self.sp = databricks.ServicePrincipal(
             f"service-principal-{sp.display_name}",
             display_name=sp.display_name,
             application_id=sp.application_id,
-            **kwargs,
+            opts=opts,
         )
 
+        self.roles = []
         for role in sp.roles:
-            databricks.ServicePrincipalRole(
+            self.roles += [databricks.ServicePrincipalRole(
                 f"service-principal-role-{sp.display_name}-{role}",
                 service_principal_id=self.sp.id,
                 role=role,
-                **kwargs,
-            )
+                opts=opts,
+            )]
 
         if not groups:
             if sp.groups:
@@ -63,6 +60,7 @@ class PulumiServicePrincipal(PulumiResourcesEngine):
             return
 
         # Group Member
+        self.group_members = []
         for g in sp.groups:
 
             # Find matching group
@@ -77,9 +75,9 @@ class PulumiServicePrincipal(PulumiResourcesEngine):
             elif isinstance(groups, dict):
                 group_id = groups[g]
 
-            databricks.GroupMember(
+            self.group_members += [databricks.GroupMember(
                 f"group-member-{sp.display_name}-{g}",
                 group_id=group_id,
                 member_id=self.sp.id,
-                **kwargs,
-            )
+                opts=opts,
+            )]